requirements.txt
flask
flask-socketio
ibm-watsonx-ai
python-dotenv
gunicorn
